code/customer-code/requirements.txt
kafka-python==1.4.7
pyshp==2.1.0
redis==3.3.11
pandas
matplotlib